moved client methods to client.go file
